requirements.txt
fastapi
uvicorn
python-dotenv
pyodbc
pypyodbc
sqlalchemy
pydantic
PyJWT
python-decouple